Prevent framebuffer modifications from switching to sys
It seems that the framebuffer data being altered in RAM should use sys, but it seems that
the sys and cfb versions of fillrect, copyarea, and imageblit are nearly the same.
Since the improvements I'm working on are based in cfb and the framebuffer still works when
continuing to use cfb with these modifications, I'm going to leave it using cfb unless
someone can convince me otherwise.